Adding support for flxlb zplit keyboard
The board is in qmk main, as well as via, but it is shy on layout options and only mentions the optional encoder in the keymap and renders. I've rounded out everything missing and think it's ready to go. This is my first ever PR; please tell me if I should be more descript or anything.