support rgb lighting for reviung rgb
The reviung34sb (formerly called reviung34) do not have rgb lighting
But the unibody one (see keymaps/reviung34_rgb) has support.

This makes the firmware quite big and all animations will not fit on the board.
Animations were not working for me, but I included in case it is working for others.